Moved some directories around to make it cleaner
