+++ node-video-crime-miner/src/interfaces/database/postgres/db.sql
