+++ node-video-crime-miner/src/interfaces/video-storage/s3Connector.ts
