+++ node-video-crime-miner/test/express-routes/api/api.labels.test.ts
@@ -1,6 +1,6 @@
 import { getMockReq, getMockRes } from '@jest-mock/express'
 import { fetchLabelDetectionJob, fetchLabelDetectionIdsForFile, fetchAllLabelDetectionForMultipleFiles, createNewLabelDetectionJob } from '../../../src/express-routes/api/api.labels'
-import * as dbLabels from '../../../src/postgres/db.labels'
+import * as dbLabels from '../../../src/interfaces/database/postgres/db.labels'
 import * as labelUtils from '../../../src/AWS Layer/Rekognition/videoLabelUtils'
 
 // FetchLabelDetectionJob function test
